tests/wpt/web-platform-tests/css/css-text-decor/text-decoration-thickness-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Decoration Test: parsing text-decoration-thickness computed values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-4/#text-decoration-width-property">
<meta name="assert" content="text-decoration-thickness computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-decoration-thickness", "auto");
test_computed_value("text-decoration-thickness", "from-font");
test_computed_value("text-decoration-thickness", "calc(10px - 8px)", "2px");
</script>
</body>
</html>
